Refactor import statements to remove unused dependencies
