+++ src/handlers/auth/auth-schemas.ts
@@ -1,4 +1,4 @@
-import { createInsertSchema, createSelectSchema } from "drizzle-zod";
+import { createInsertSchema } from "drizzle-zod";
 import { z } from "zod";
 
 import * as schema from "@/db/schema";
@@ -9,14 +9,8 @@ export const signupUserSchema = createInsertSchema(schema.users).omit({
   updatedAt: true,
 });
 export const signinUserSchema = signupUserSchema.pick({ email: true, password: true });
-export const updateUserSchema = signupUserSchema.partial();
-export const foundUserSchema = createSelectSchema(schema.users);
 
-export type TUserToCreate = z.infer<typeof signupUserSchema>;
-export type TUserToUpdate = z.infer<typeof updateUserSchema>;
-export type TCandidateUser = z.infer<typeof signinUserSchema>;
-export type TFoundUserResult = z.infer<typeof foundUserSchema>;
-
+// Response serialization schemas
 export const authResponseSchema = z.object({
   message: z.string(),
 });
